[CLI] Added --chef_repo_path option
This will override the chef_repo_paths option in the config. When
specified, the directory is assumed to contain the entire chef repo
you want.

This option is put in place so that testing migrations using
`knife download` and `knife upload` is easier. If "override everything" is
a feature request you require, then please file a bug report